PoC Pong Slack bot
This is a very basic Proof-of-Concept for a Slack bot in GitLab. Main
focus of this commit just to get something running and iterate from
this point on.

If you'd like to test this, you'll need a Slack team and a bot with a
token/key. In case you're a GitLab team member, I can supply these to
you.

[ci skip]