Adopt a new status 'success_with_warnings' for pipeline, and:
HasStatus.status (jobs collection). This would be used for showing
a warning icon for pipelines/stages which contains some jobs which
were failed or cancelled but it's allowed.

We don't do this for individual builds because we don't want to
lose information.

TODO: Add some tests for this particular case.