So that we could show success_with_warnings for:
a particular commit. It's a bit weird that commit status
is actually the status from a collection of pipelines,
while CommitStatus.status is a status from a collection of
builds/jobs, which is also the status of a pipeline/stage.

We have a lot of confusing names for now, unfortunately.