Add GitlabSpecListener spec support file to monitor passes and failures
The plan is to run the suite in a known-good state, recording the passes
in a common format. Then we'll change all of the `create(:project)`
calls to `create(:empty_project)` and record the passes _and_ the
failures. After removing the list of failures in the second run from the
list of successes in the first one, anything still in the "success" list
can stay.

See https://gitlab.com/gitlab-org/gitlab-ce/issues/24899#note_19062181