Allow object store for Job Artifacts
Given that storage of artifacts on disk is not the most scalable
solution, object storage, e.g. S3 should be supported. The change
requires a new key in the `gitlab.yml`, `object_store` under artifacts.
If this is set to true, fog will be used for storage instead the local
FS.

If credentials are incorrect, we don't catch this yet.