Add many foreign keys to the projects table
This removes the need for relying on Rails' "dependent" option for data
removal, which is _incredibly_ slow (even when using :delete_all) when
deleting large amounts of data. This also ensures data consistency is
enforced on DB level and not on application level (something Rails is
really bad at).

This also updates the Project model to take care of removing artifacts
and the likes as this is not handled by PostgreSQL when it removes data
in a cascading fashion.