Use rspec-set to speed up examples
The gem, rspec-set, persists the object only once per context in which
they are called. On the files touched this shaved off a minute, out of
the 4 total. Keep in mind I only touched top level, to check if this
worked as expected.

Especially cascading effect seem promising. Top level `before` blocks
can probably be a `before(:all)` now, which saves time too, when for
example adding members to a project.

Stats of the files I touched:
- before: 4:26 minutes
- after 3:09 minutes